Commented Quaternion and Vector3 + Removed unused RollNickGier

+++ MyMathLib/Vector3.cs
@@ -2,39 +2,54 @@
 using System.Text;
 
 namespace MyMathLib {
+    /// <summary>A standard vector class using float components</summary>
     public class Vector3 {
+        /// <summary>This vectors components</summary>
         private readonly float[] vector;
 
+        /// <summary>Construct a vector initialized with (0,0,0)</summary>
         public Vector3() {
             this.vector = new float[3];
         }
 
+        /// <summary>
+        /// Construct a vector with 3 components
+        /// </summary>
         public Vector3(float x, float y, float z) {
             this.vector = new[] { x, y, z };
         }
 
+        /// <summary>
+        /// Access the vectors components trough this indexer
+        /// </summary>
+        /// <param name="i"></param>
         public float this[int i] {
             get => vector[i];
             set => vector[i] = value;
         }
 
+        /// <summary>Access the vectors x component</summary>
         public float x {
             get => vector[0];
             set => vector[0] = value;
         }
 
+        /// <summary>Access the vectors y component</summary>
         public float y {
             get => vector[1];
             set => vector[1] = value;
         }
 
+        /// <summary>Access the vectors z component</summary>
         public float z {
             get => vector[2];
             set => vector[2] = value;
         }
 
+        /// <summary>Calculate the length of this vector</summary>
         public float Length => (float)Math.Sqrt(Math.Pow(x,2) + Math.Pow(y,2) + Math.Pow(z,2));
         
+        /// <summary>Get this vector normalized (returns a copy)</summary>
         public Vector3 Normalized => this / this.Length;
 
         public static Vector3 operator +(Vector3 a, Vector3 b) {
@@ -81,8 +96,11 @@ namespace MyMathLib {
             return ret;
         }
 
-        public static double Dot(Vector3 a, Vector3 b) {
-            var ret = 0d;
+        /// <summary>
+        /// Calculates the dot product of two vectors
+        /// </summary>
+        public static float Dot(Vector3 a, Vector3 b) {
+            var ret = 0f;
             for (var i = 0; i < 3; i++) {
                 ret += a[i] * b[i];
             }
@@ -90,6 +108,9 @@ namespace MyMathLib {
             return ret;
         }
 
+        /// <summary>
+        /// Caclulates the cross product of two vectors. (returns a copy)
+        /// </summary>
         public static Vector3 Cross(Vector3 a, Vector3 b) {
             return new Vector3{
                 x = a.y * b.z - a.z * b.y,
